refactor: extract openai calls to wrapper and facade
