+++ src/lib/ai/answer.ts
@@ -1,59 +1,14 @@
 import {getOpenAIAPI} from './config'
-import {createEmbedding} from './createEmbedding';
 import {callWithRetry} from './callWithRetry'
 import {logger} from './logger'
 import {QnaStorage} from 'src/lib/ai/largeDocQna'
-
-const completionConfig = {
-  temperature: 0,
-  max_tokens: 256,
-  top_p: 1,
-  frequency_penalty: 0,
-  presence_penalty: 0,
-}
-
-const defaultPrompt = "Answer the question as truthfully as possible using the provided text, and if the answer is not contained within the text below, say \"I don't know\"\n\n"
-
-async function answerMe(context: string, userPrompt: string, initPrompt:string = defaultPrompt) {
-  const prompt = initPrompt
-    + 'Context:\n' + context + '\n\n'
-    + 'Q: ' + userPrompt + '\nA: ';
-
-  logger.debug(prompt)
-
-  const response = await callWithRetry(() => getOpenAIAPI().createChatCompletion({
-    ...completionConfig,
-    model: 'gpt-3.5-turbo',
-    messages:[
-      {role: 'system', content: context},
-      {role: 'user', content: initPrompt},
-      {role: 'user', content: userPrompt}
-    ],
-    // user: TODO: for tracking purposes
-  }));
-
-  // const response = await openai.createCompletion({
-  //     ...completionConfig,
-  //     // stop: ["Human:", "AI:", "Olea Bot:"],
-  //     model: Config.engineId,
-  //     prompt,
-  //
-  //     // user: TODO: for tracking purposes
-  //
-  //
-  // });
-
-  let res: string = response.data!.choices![0].message!.content ?? '';
-  res = ((/^\s*$/g).test(res) ? 'Failed' : res);
-  // Could potentially also detect "I don't know" and return undefined in that case
-  return res
-}
+import {answerMe, createEmbedding} from 'src/lib/ai/openaiFacade'
 
 export async function performQna(question:string, storage: QnaStorage, prompt?: string): Promise<string|undefined> {
   // const loader = loaderFromOpts(opts)
   const data: { keys: string[], embeds: number[][] } = await storage.readEmbeds()
   // JSON.parse(await readFile(embedsFile, 'utf-8'));
-  const qryEmbed = await createEmbedding(question)
+  const qryEmbed = await createEmbedding({input:question})
   // Find most similar segment index to all other inputs
   const similarities = findMostSimilarSegments(qryEmbed, data.embeds)
 
@@ -74,7 +29,7 @@ export async function performQna(question:string, storage: QnaStorage, prompt?: 
   logger.info(`best inputs: ${contexts.length} snippets`)
   if (contexts.length>0) {
     const bestInput = contexts.join('\n')
-    const answer = await answerMe(bestInput, question, prompt)
+    const answer = await answerMe({context: bestInput, userPrompt: question, initPrompt: prompt})
     return answer
   } else {
     return undefined
